267 - fixed circular import
